Refactor BBPE based Tokenizer
