+++ cmake/externals/opencv.cmake
@@ -162,5 +162,6 @@ if (CMAKE_SYSTEM_NAME MATCHES "Windows")
     list(APPEND opencv_projs opencv_highgui_plugins opencv_videoio_plugins)
     foreach(p ${opencv_projs})
         set_target_properties(${p} PROPERTIES FOLDER "externals/opencv")
+        set_target_properties(${p} PROPERTIES EXCLUDE_FROM_ALL TRUE EXCLUDE_FROM_DEFAULT_BUILD TRUE)
     endforeach()
 endif()
