--- operators/tokenizer/gpt2_tokenizer.hpp
@@ -1,18 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-#pragma once
-#include "bpe_tokenizer.hpp"
-
-struct KernelBpeTokenizer : BaseKernel {
-  KernelBpeTokenizer(const OrtApi& api, const OrtKernelInfo& info);
-  void Compute(const ortc::Tensor<std::string>& input,
-               ortc::Tensor<int64_t>& tokenize_output,
-               std::optional<ortc::Tensor<int64_t>*> attention_mask) const;
-
- private:
-  std::vector<int64_t> Tokenize(const ustring& input, int64_t max_length) const;
-
-  int64_t padding_length_;
-  std::shared_ptr<VocabData> bbpe_tokenizer_;
-};
