+++ test/test_robertatok.py
@@ -80,27 +80,16 @@ class TestRobertaTokenizer(unittest.TestCase):
         sess = _ort.InferenceSession(model.SerializeToString(), so, providers=['CPUExecutionProvider'])
         input_text = np.array(test_sentence)
         input_ids, attention_mask, offset_mapping = sess.run(None, {'string_input': input_text})
-        print("\nTest Sentence: " + str(test_sentence))
-        print("\nInput IDs: " + str(input_ids))
-        print("Attention Mask: " + str(attention_mask))
-        # Reformat offset mapping from 3d array to 2d array of tuples before printing for readability
-        reformatted_offset_mapping = nlr.unstructured_to_structured(np.array(offset_mapping)).astype('O')
-        print("Offset Mapping: " + str(reformatted_offset_mapping))
         roberta_out = self.tokenizer(test_sentence, return_offsets_mapping=True)
         expect_input_ids = roberta_out['input_ids']
         expect_attention_mask = roberta_out['attention_mask']
         expect_offset_mapping = roberta_out['offset_mapping']
-        print("\nExpected Input IDs: " + str(expect_input_ids))
-        print("Expected Attention Mask: " + str(expect_attention_mask))
-        print("Expected Offset Mapping: " + str(expect_offset_mapping) + "\n")
         np.testing.assert_array_equal(expect_input_ids, input_ids)
         np.testing.assert_array_equal(expect_attention_mask, attention_mask)
         np.testing.assert_array_equal(expect_offset_mapping, offset_mapping)
 
-        del sess
-        del so
-
     def test_tokenizer(self):
+        self._run_tokenizer([" ", "\n"])
         self._run_tokenizer(["I can feel the magic, can you?"])
         self._run_tokenizer(["Hey Cortana"])
         self._run_tokenizer(["lower newer"])
